+++ CRM/CivirulesCronTrigger/Form/NextContributionDate.php
@@ -13,8 +13,8 @@ class CRM_CivirulesCronTrigger_Form_NextContributionDate extends CRM_CivirulesTr
    */
   public function buildQuickForm() {
     $this->add('hidden', 'rule_id');
-    $this->add('select', 'interval_unit', ts('Interval'), CRM_CivirulesCronTrigger_MembershipEndDate::intervals(), TRUE);
-    $this->add('text', 'interval', ts('Interval'), TRUE);
+    $this->add('select', 'interval_unit', ts('Interval'), CRM_CivirulesCronTrigger_NextContributionDate::intervals(), TRUE);
+    $this->add('text', 'interval', ts('Interval'), [], TRUE);
     $this->addRule('interval', ts('Interval should be a numeric value'), 'numeric');
     $this->addButtons([
       ['type' => 'next', 'name' => ts('Save'), 'isDefault' => TRUE,],